refactor(tui): colocate categorical theme defaults
