+++ packages/tui/src/theme/v2/schema.ts
@@ -36,14 +36,6 @@ const ColorValue = Schema.Union([
 
 export const HueName = Schema.Union([BaseHue, HueAlias])
 export type HueName = Schema.Schema.Type<typeof HueName>
-export const DEFAULT_CATEGORICAL = [
-  "blue",
-  "purple",
-  "green",
-  "orange",
-  "red",
-  "cyan",
-] as const satisfies readonly HueName[]
 export const CategoricalDefinition = Schema.Array(HueName).check(Schema.isMinLength(1))
 export type CategoricalDefinition = Schema.Schema.Type<typeof CategoricalDefinition>
 const HueColorValue = Schema.Union([HexColor, Schema.TemplateLiteral(["$hue.", HueName, ".", HueStep])])
